refactor: move chat prompt sanitization into codersdk
Exports SanitizePromptText (with its helpers) from codersdk instead of
coderd/x/chatd so API consumers can sanitize prompt text exactly the
way the server stores it. terraform-provider-coderd needs this for the
coderd_chat_system_prompt resource, whose semantic equality must match
the server's sanitization to avoid false drift; until now it carried a
mirrored copy that could rot.

Pure move plus callsite updates: no behavior change, tests move along
with it.
